Split platform services creation to allow logger to be passed in for network services; move browser-specific utilities for network to browser package
