+++ packages/browser/src/services/injectSendBeaconRequest.js
