+++ packages/browser/src/services/injectSendFetchRequest.js
