+++ packages/browser/test/unit/specs/services/injectSendFetchRequest.spec.js
@@ -11,7 +11,7 @@ governing permissions and limitations under the License.
 */
 
 import { vi, describe, it, expect } from "vitest";
-import injectSendFetchRequest from "../../../../../../src/core/network/requestMethods/injectSendFetchRequest.js";
+import injectSendFetchRequest from "../../../../src/services/injectSendFetchRequest.js";
 
 describe("injectSendFetchRequest", () => {
   it("resolves returned promise upon network success", () => {
