+++ packages/core/src/core/index.js
@@ -21,7 +21,6 @@ import createLogController from "./createLogController.js";
 import createLifecycle from "./createLifecycle.js";
 import createComponentRegistry from "./createComponentRegistry.js";
 import injectSendNetworkRequest from "./network/injectSendNetworkRequest.js";
-import injectSendBeaconRequest from "./network/requestMethods/injectSendBeaconRequest.js";
 import injectExtractEdgeInfo from "./edgeNetwork/injectExtractEdgeInfo.js";
 import createIdentity from "./identity/createIdentity.js";
 import createConsent from "./consent/createConsent.js";
@@ -82,6 +81,7 @@ export const createExecuteCommand = ({
   });
 
   const { setDebugEnabled, logger, createComponentLogger } = logController;
+  const network = platformServices.createNetworkService(logger);
 
   const parsedQueryString = queryString.parse(
     platformServices.globals.getLocationSearch(),
@@ -144,14 +144,10 @@ export const createExecuteCommand = ({
       dateProvider: () => new Date(),
     });
 
-    const { sendFetchRequest, sendBeacon } = platformServices.network;
-    const sendBeaconRequest = sendBeacon
-      ? injectSendBeaconRequest({ sendBeacon, sendFetchRequest, logger })
-      : sendFetchRequest;
     const sendNetworkRequest = injectSendNetworkRequest({
       logger,
-      sendFetchRequest,
-      sendBeaconRequest,
+      sendFetchRequest: network.sendFetchRequest,
+      sendBeaconRequest: network.sendBeaconRequest,
       isRequestRetryable,
       getRequestRetryDelay,
     });
